requirements.txt
peewee==3.13.3
futures==3.1.1
grpcio==1.32.0
protobuf==3.13.0
grpcio-tools==1.32.0
python-consul==1.1.0
redis==3.4.1
pymysql==0.9.3
ronglian_sms_sdk==1.0
